+++ src/loaders/support/LoaderWorkerSupport.js
@@ -93,7 +93,7 @@ THREE.LoaderSupport.WorkerRunnerRefImpl = (function () {
  */
 THREE.LoaderSupport.WorkerSupport = (function () {
 
-	var WORKER_SUPPORT_VERSION = '1.0.0';
+	var WORKER_SUPPORT_VERSION = '1.1.0';
 
 	var Validator = THREE.LoaderSupport.Validator;